test(desktop): split sidebar local slot coverage
